feat: exception tidy up
