+++ NeverBounceSDK/Exceptions/BadReferrerException.cs
@@ -1,25 +1,24 @@
-﻿using System.Runtime.Serialization;
-
-namespace NeverBounce.Exceptions;
+﻿namespace NeverBounce.Exceptions;
+using System.Runtime.Serialization;
 
 [Serializable]
-public class GeneralException : Exception
+public class BadReferrerException : Exception
 {
-    public GeneralException()
+    public BadReferrerException()
     {
     }
 
-    public GeneralException(string message) : base(message)
+    public BadReferrerException(string message) : base(message)
     {
     }
 
-    public GeneralException(string message, Exception inner) : base(message, inner)
+    public BadReferrerException(string message, Exception inner) : base(message, inner)
     {
     }
 
     // A constructor is needed for serialization when an
     // exception propagates from a remoting server to the client. 
-    protected GeneralException(SerializationInfo info,
+    protected BadReferrerException(SerializationInfo info,
         StreamingContext context)
     {
     }
